refactor: Rename project container component as its not a section itself
